[#86] API関係の処理をpresenterからmodelに移動した
